Bity logo on recap (#35)
* rename routes

* Add bity logo on recap page

* Add bity link
