+++ packages/widget/package.json
@@ -5,7 +5,7 @@
     "type": "git",
     "url": "https://github.com/pakokrew/mooni.git"
   },
-  "version": "0.1.4",
+  "version": "0.2.0",
   "license": "GPL-3.0",
   "main": "build/index.cjs.js",
   "module": "build/index.es.js",
